chore: assert with miniflare
