+++ clients/algoliasearch-client-javascript/packages/requester-browser-xhr/index.ts
@@ -1,2 +1,1 @@
 export * from './src/createXhrRequester';
-export * from './src/echoRequester';
