+++ clients/algoliasearch-client-javascript/packages/requester-node-http/index.ts
@@ -1,2 +1,1 @@
-export * from './src/echoRequester';
 export * from './src/createHttpRequester';
