+++ templates/javascript/clients/algoliasearch/builds/definition.mustache
@@ -21,9 +21,9 @@ import type {
 export * from './models';
 
 export type Algoliasearch = SearchClient & {
-    initRecommend: (initOptions: InitClientOptions)=> RecommendClient;
-    initAnalytics: (initOptions: InitClientOptions & InitClientRegion<AnalyticsRegion>)=> AnalyticsClient;
-    initAbtesting: (initOptions: InitClientOptions & InitClientRegion<AbtestingRegion>)=> AbtestingClient;
+    initRecommend: (initOptions?: InitClientOptions)=> RecommendClient;
+    initAnalytics: (initOptions?: InitClientOptions & InitClientRegion<AnalyticsRegion>)=> AnalyticsClient;
+    initAbtesting: (initOptions?: InitClientOptions & InitClientRegion<AbtestingRegion>)=> AbtestingClient;
     initPersonalization: (initOptions: InitClientOptions & Required<InitClientRegion<PersonalizationRegion>>)=> PersonalizationClient;
 };
